requirements.txt
glfw==2.5.5
keyboard==0.13.5
numpy==1.24.2
pygame==2.1.2
PyOpenGL==3.1.6
scipy==1.9.1
